edit Brooklyn license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit updates the LICENSE file so that it begins with the full
text of the Apache license (including the Appendix). The text that
preceded the beginning of the license text has been transferred to a
new "License" section in the README.

Collectively, these changes allow Licensee to successfully identify the
license type of Brooklyn as Apache 2.0.

REFS: BROOKLYN-593
Signed-off-by: Andrea Kao <eirinikos@gmail.com>